Create new element with mass as float, not string
forcefield.registerAtomType() was creating a new element with the mass
that ParmEd was passing it, which is a string. The mass was converted to
a floating point number in the function, but that was not what was
passed to simtk.openmm.app.element.Element().  This was causing errors
in the hoomdxml writer when trying to divide by the reference mass
(since hoomd uses reduced units).