Begin adding atom types for Trappe alkenes
Added atom types and bond parameters for Trappe alkenes. Stopped as
I'm not sure how to handle chirality in the dihedrals. Renamed atom
types based on hybridization, although I think it might be best to
eventually switch to the convention that is used for OPLS (e.g.
trappe_1, trappe_2, etc.).