Quiet output for missing atom types
Previously, when applying a forcefield with many missing atom types
in the forcefield XML, Foyer would output as a warning all the
missing atom types to the user upon successfully atom typing.

For the larger forcefield files (oplsaa.xml for example), there are currently
~750 missing atom types that do not have a smarts string. This led to cumbersome
warning messages.

Now, the default behavior for missing atom types is to still warn the user, but
condense the information to ~2 lines of output.

If the user wants to know which atom types are missing a smarts string,
a debug flag can be passed during the application of the forcefield to
revert to the previous behavior.