Change "OS" oxygen atoms in silica mol2 file to O
With the latest updates to mBuild in version 0.9.3, the atom names when
being converted to various file formats or parmed structures now undergo
a different process when their elements are being inferred.

Previously, an element name like `OS`, like in this silica.mol2 file
would evaulate to O, since `OS` is not an element.

This usually caused more issues though, because certain elements like
`SI` would be inferred as sulfur.

This now causes issues with some of our tests that rely on this silica
mol2 file, as OS is now being discovered as osmium.

The OS elements in the mol2 file have been renamed to O.