Increased support of amide atomtyping with OPLS-AA.
Added definitions, descriptions, and doi for three atom types related to amides. I was running into issues using foyer to atomtype NMP and NMA for a small simulation.
All atom types are from Jorgensen et al OPLS-AA paper doi: 10.1021/ja9621760

The next step is to test these atomtypes with specific examples annd include overrides.

opls_237: describes the N in a primary amide such as carboxymide
opls_238: describes the N in a secondary amide such as NMA
opls_245: describes a secondary C bonded to an amide N such as in NMP

I also edited the description to opls_243 to specifically indicate that the carbon is primary.

opls_243: describes a primary C bonded to an amide N

This work-around allows me to atomtype NMP. Adding this should allow the modelling of proteins and any other molecules containing amide bonds. A more extensive search of the Jorgensen et al paper is required to ensure that all described amide parameters are supported.