Added more amide support.
Added definitions, descriptions, overrides, and doi for the following atomtypes related to amides.

237: primary amide N
238: secondary amide N
240: H attached to primary amide
241: H attached to secondary amide
242: methyl C attached to secondary amide
245: methyl C attached to tertiary amide

I tested these with acetamide, N-methylacetamide, and NMP. Before these were added, these molecules could not be atomtyped and returned an error.

SMARTS strings still need to be added for the C in Gly (opls_244) and I cannot locate anything for the C in a RCH2 attached to a secondary amide such as in N-ethylacetamide. This is most likely from a different paper.