include setuptools at build and runtime for conda
To ensure that we have our plugins supported properly when installed
from conda, it appears that a few options are necessary to ensure that
is the case.

For the case of setuptools entry_points, the conda-build documentation
recommends that we do include the build option: preserve_egg_dir

https://docs.conda.io/projects/conda-build/en/latest/user-guide/recipes/build-without-recipe.html#preserve-egg-directory

Setuptools are also now build and runtime dependencies.