Add get_available_forcefield_loaders (#388)
This PR adds a get_available_forcefield_loaders function, which returns a list of the functions returning Forcefield objects as discovered by the plugin system. I believe there is no other way to systematically access these function in the current API, i.e. from inside Foyer and not the package that plugs into it. In the past I have either called entry points I had in short-term memory or tab-completed foyer.forcefields.load_ in an interpreter.