Include git in docker image to clone repo
With the new additions of the conda env yml files to our repo, this
inevitably added some unforseen issues with our release pipeline.

This fixes the case where our docker images would not build, since the
environment-dev.yml file has a `pip:` section that needs to clone a
repository using `git`. `git` is not installed by default in the alpine
docker image, and has been added to the `RUN` command.