Include pre-commit configuration files
Foyer currently lacks an automated way to enforce a style guide and
basic docstring formatting that is easily distributed to other
developers.

An example of this is mBuild, which now uses the `pre-commit` framework
to automate this during pull-requests as well as an installable git
commit hook.
This commit adds the initial configuration files for pre-commit setup.