Allow pre-commit to autoupdate prs
Previously, when we introduced pre-commit, it required users to manually
run the pre-commit hook locally.

Now, these pre-commit features can be run in the cloud now, which can be
useful for users that might not want to install the required packages.
This introduces the yaml configuration necessary to autoformat PRs
according to our current configuration.