Change ff load logic to make options consistent
When a forcefield file is provided by using the internal name:
`foyer.Forcefield(name='oplsaa)`, the final file processing would happen
in a separate part of the logic and use variables that were only
accessed if the user used `forcefield_files` in the `Forcefield`
constructor method.

Now the same final processing occurs for all input types to load the
forcefields. All tests currently pass with this fix in place.