Include detailed tests for forcefield metadata (#435)
* Include detailed tests for forcefield metadata

Previously, the tests in test_forcefield.py were assumed to test that
forcefield metadata like `version`, `name`, etc. were properly loaded
into the `Forcefield` object.

This is only partially true, as there are cases where the files either
go out of scope before that information is gathered, etc. This PR
includes additional tests to ensure that this information is not being
lost prematurely. Big thanks to: @mattwthompson, @ahy3nz for discovering
this issue, creating MWE's for some of these cases, and taking my
investigation further and most likely pinpointing the issue.

* Change ff load logic to make options consistent

When a forcefield file is provided by using the internal name:
`foyer.Forcefield(name='oplsaa)`, the final file processing would happen
in a separate part of the logic and use variables that were only
accessed if the user used `forcefield_files` in the `Forcefield`
constructor method.

Now the same final processing occurs for all input types to load the
forcefields. All tests currently pass with this fix in place.

* Remove named temp files after final parsing