Include combining_rule for trappe-ua.xml
After the addition of the `combining_rule` forcefield XML metadata, we
added the `geometric` mixing rule to oplsaa.xml, but neglected to
include that information for our TraPPE xml.

This includes the `Lorentz-Berethlot` mixing rule as metadata as well as
bumping the version number for the trappe-ua.xml file.

A minor spelling mistake was also fixed.